+++ modules/contracts/src.sol/lib/LibUtils.sol
@@ -13,4 +13,16 @@ library LibUtils {
         return x <= y ? y : x;
     }
 
+    // If success is false, reverts and passes on the revert string.
+    function revertIfCallFailed(bool success, bytes memory returnData)
+        internal
+        pure
+    {
+        if (!success) {
+            assembly {
+                revert(add(returnData, 0x20), mload(returnData))
+            }
+        }
+    }
+
 }